Use maven shade plugin
I was having problems with missing spring handlers. This can sometimes happen when multiple spring dependencies contain properties files with the same name but different contents: building the final jar can result in the maven assembly plugin overwriting one spring properties with another of the same name.
The maven shade plugin helps resolve this.
For more about this problem, see: https://stackoverflow.com/q/19615698 for an example of the problem and solution.